requirements.txt
amqp==2.5.0
billiard==3.5.0.5
celery==4.2.1
Django==2.2.2
django-cors-headers==3.0.2
django-redis==4.10.0
djangorestframework==3.9.4
django-prices==2.1.0
kombu==4.6.3
psycopg2-binary==2.7.5
pytz==2019.1
redis==3.2.0
sqlparse==0.3.0
vine==1.3.0
Pillow==6.2.2
django-ckeditor==5.8.0
django-js-asset==1.2.2
pandas==0.25.3
gunicorn==20.0.4
requests==2.22.0